Recommendation to solve issue in order to run on Netlify
Diagnosis: The build failed due to an error related to a missing dependency in babel-preset-react-app, specifically the @babel/plugin-proposal-private-property-in-object.

Solution: To resolve this issue, you need to add @babel/plugin-proposal-private-property-in-object to your devDependencies.

Before adding the dependency, verify that the package is in your package.json file or was committed to the repository. Once verified, add the package to your devDependencies by running the following command:

npm install --save-dev @babel/plugin-proposal-private-property-in-object

After adding the dependency, commit the changes to your repository and trigger a new build to see if the issue is resolved.